更新2.14.0

+++ book/src/main/kotlin/book/util/help/RuleBigDataHelp.kt
@@ -44,8 +44,9 @@ object RuleBigDataHelp {
     }
 
     fun putBookVariable(bookUrl: String,userid :String, key: String, value: String?) {
-        if(userid.isEmpty()) return
+        if(userid.isEmpty() || bookUrl.isEmpty()) return
         val md5BookUrl = MD5Utils.md5Encode(bookUrl)
+        //println("putBookVariable $bookUrl,$userid,$key,$value,$md5BookUrl")
         val md5Key = MD5Utils.md5Encode(key)
         if (value == null) {
             FileUtils.delete(FileUtils.getPath(bookData,userid, md5BookUrl, "$md5Key.txt"), true)
@@ -60,7 +61,7 @@ object RuleBigDataHelp {
     }
 
     fun getBookVariable(bookUrl: String,userid :String, key: String?): String? {
-        if(userid.isEmpty()) return null
+        if(userid.isEmpty() || bookUrl.isEmpty()) return null
         val md5BookUrl = MD5Utils.md5Encode(bookUrl)
         val md5Key = MD5Utils.md5Encode(key)
         val file = File(FileUtils.getPath(bookData,userid, md5BookUrl, "$md5Key.txt"))
@@ -72,7 +73,7 @@ object RuleBigDataHelp {
 
 
     fun putChapterVariable(bookUrl: String,userid :String, chapterUrl: String, key: String, value: String?) {
-        if(userid.isEmpty()) return
+        if(userid.isEmpty() || bookUrl.isEmpty()) return
         val md5BookUrl = MD5Utils.md5Encode(bookUrl)
         val md5ChapterUrl = MD5Utils.md5Encode(chapterUrl)
         val md5Key = MD5Utils.md5Encode(key)
@@ -90,7 +91,7 @@ object RuleBigDataHelp {
     }
 
     fun getChapterVariable(bookUrl: String,userid :String, chapterUrl: String, key: String): String? {
-        if(userid.isEmpty()) return null
+        if(userid.isEmpty() || bookUrl.isEmpty()) return null
         val md5BookUrl = MD5Utils.md5Encode(bookUrl)
         val md5ChapterUrl = MD5Utils.md5Encode(chapterUrl)
         val md5Key = MD5Utils.md5Encode(key)
